Move all installation code to a new installation component
