+++ src/NServiceBus.Core/EndpointCreator.cs
@@ -7,7 +7,6 @@ namespace NServiceBus
     using System.Threading.Tasks;
     using Features;
     using Hosting.Helpers;
-    using Installation;
     using MessageInterfaces;
     using MessageInterfaces.MessageMapper.Reflection;
     using ObjectBuilder;
@@ -110,12 +109,9 @@ namespace NServiceBus
 
             pipelineComponent.AddRootContextItem<IEventAggregator>(eventAggregator);
 
-            var shouldRunInstallers = settings.GetOrDefault<bool>("Installers.Enable");
+            installationComponent = new InstallationComponent(settings);
 
-            if (shouldRunInstallers)
-            {
-                RegisterInstallers(concreteTypes);
-            }
+            installationComponent.Initialize(concreteTypes, containerComponent, receiveComponent);
 
             settings.AddStartupDiagnosticsSection("Endpoint",
                 new
@@ -137,46 +133,9 @@ namespace NServiceBus
 
         public async Task<IStartableEndpoint> CreateStartableEndpoint()
         {
-            var shouldRunInstallers = settings.GetOrDefault<bool>("Installers.Enable");
-
-            if (shouldRunInstallers)
-            {
-                var username = GetInstallationUserName();
-
-                if (settings.CreateQueues())
-                {
-                    await receiveComponent.CreateQueuesIfNecessary(queueBindings, username).ConfigureAwait(false);
-                }
-
-                await RunInstallers(containerComponent.Builder, username).ConfigureAwait(false);
-            }
+            await installationComponent.Start().ConfigureAwait(false);
 
             return new StartableEndpoint(settings, containerComponent, featureComponent, transportInfrastructure, receiveComponent, criticalError, pipelineComponent, recoverabilityComponent);
-        }
-
-        async Task RunInstallers(IBuilder builder, string username)
-        {
-            foreach (var installer in builder.BuildAll<INeedToInstallSomething>())
-            {
-                await installer.Install(username).ConfigureAwait(false);
-            }
-        }
-
-        string GetInstallationUserName()
-        {
-            if (!settings.TryGet("Installers.UserName", out string userName))
-            {
-                if (Environment.OSVersion.Platform == PlatformID.Win32NT)
-                {
-                    userName = $"{Environment.UserDomainName}\\{Environment.UserName}";
-                }
-                else
-                {
-                    userName = Environment.UserName;
-                }
-            }
-
-            return userName;
         }
 
         TransportInfrastructure InitializeTransportComponent()
@@ -282,14 +241,6 @@ namespace NServiceBus
             settings.TryGet("onCriticalErrorAction", out Func<ICriticalErrorContext, Task> errorAction);
             criticalError = new CriticalError(errorAction);
             containerComponent.ContainerConfiguration.RegisterSingleton(criticalError);
-        }
-
-        void RegisterInstallers(IEnumerable<Type> concreteTypes)
-        {
-            foreach (var installerType in concreteTypes.Where(t => IsINeedToInstallSomething(t)))
-            {
-                containerComponent.ContainerConfiguration.ConfigureComponent(installerType, DependencyLifecycle.InstancePerCall);
-            }
         }
 
         static List<Type> PerformAssemblyScanning(EndpointConfiguration endpointConfiguration)
@@ -400,8 +351,6 @@ namespace NServiceBus
 
         static bool HasDefaultConstructor(Type type) => type.GetConstructor(Type.EmptyTypes) != null;
 
-        static bool IsINeedToInstallSomething(Type t) => typeof(INeedToInstallSomething).IsAssignableFrom(t);
-
         PipelineComponent pipelineComponent;
         SettingsHolder settings;
         ContainerComponent containerComponent;
@@ -411,5 +360,6 @@ namespace NServiceBus
         QueueBindings queueBindings;
         ReceiveComponent receiveComponent;
         RecoverabilityComponent recoverabilityComponent;
+        InstallationComponent installationComponent;
     }
 }
